move protobuf marshaler to examples
