+++ _examples/marshalers/protobuf/message.proto
