+++ go.mod
@@ -8,7 +8,6 @@ require (
 	github.com/nats-io/nats.go v1.20.0
 	github.com/pkg/errors v0.9.1
 	github.com/stretchr/testify v1.7.1
-	google.golang.org/protobuf v1.28.0
 )
 
 require (
