+++ pkg/nats/factory.go
@@ -8,8 +8,6 @@ func GetMarshaler(format string) MarshalerUnmarshaler {
 	switch strings.ToLower(format) {
 	case "nats-core":
 		return &NATSMarshaler{}
-	case "proto":
-		return &PBMarshaler{}
 	case "json":
 		return &JSONMarshaler{}
 	default:
